improved masking for `bragg_peaks_persistence` (#43)
* updated masking for `bragg_peaks_persistence`

In the function `bragg_peaks_persistence` from @trbritt the hard edges of the mask are resulting in peaks being detected at those edges. I propose to scan the full image for peaks and throw away the peaks that are in the masked area later. Resulted in improved results on my end.
`autocenter` is still done with the mask in place.

* Update CHANGELOG.rst

* Update indexing.py

* version bump